// superset-frontend/plugins/plugin-chart-custom-gauge/src/types.ts
export type Refs = {
    echartRef?: Ref<EchartsHandler>;
    divRef?: RefObject<HTMLDivElement>;
};

export interface EchartsProps {
    height: number;
    width: number;
    echartOptions: EChartsCoreOption;
    eventHandlers?: EventHandlers;
    zrEventHandlers?: EventHandlers;
    selectedValues?: Record<number, string>;
    forceClear?: boolean;
    refs: Refs;
    vizType?: string;
}

export interface EchartsHandler {
    getEchartInstance: () => EChartsType | undefined;
}